+++ src/Patient/Patient.js
@@ -140,10 +140,10 @@ class Patient extends Component {
                             tooltip: 'Remove All Selected Users',
                             icon: 'delete',
                             onClick: (evt, data) => {
+                                if (window.confirm("האם אתה בטוח?") == true) {
                                 data.forEach(data1 => {
                                     axios.delete(`patients/` + data1.id + '.json')
-                                });
-
+                                })};
                             }
